+++ tsc/src/level/level.hpp
@@ -109,11 +109,11 @@ namespace TSC {
         /* handle joystick button down event
          * returns true if processed
         */
-        bool Joy_Button_Down(Uint8 button);
+        bool Joy_Button_Down(unsigned int button);
         /* handle joystick button up event
          * returns true if processed
         */
-        bool Joy_Button_Up(Uint8 button);
+        bool Joy_Button_Up(unsigned int button);
 
         /* Return the current Music filename relative to the music/
          * directory.
